Matcher: Common module for resource matchers
